THE GALAXY MILESTONES (part 1)
Branching this one because of how big of a change it is, and how volitile it is in terms of what could happen to it in the future.

I basically ripped off Aarex's "Quantum speedruns" to  create a new tab in Galaxy. it is definitely not set to be merged, do not even try.